+++ pom.xml
@@ -449,7 +449,6 @@
                             <excludes>
                                 <exclude>**/*ContractTest.java</exclude>
                                 <exclude>**/*ProviderContractTestSuite.java</exclude>
-                                <exclude>**/*ConsumerContractTestSuite.java</exclude>
                             </excludes>
                         </configuration>
                     </plugin>
@@ -471,27 +470,6 @@
                         <configuration>
                             <includes>
                                 <include>**/*ProviderContractTestSuite.java</include>
-                            </includes>
-                        </configuration>
-                    </plugin>
-                </plugins>
-            </build>
-        </profile>
-        <profile>
-            <id>consumer-contract-tests</id>
-            <activation>
-                <property>
-                    <name>runConsumerContractTests</name>
-                </property>
-            </activation>
-            <build>
-                <plugins>
-                    <plugin>
-                        <artifactId>maven-surefire-plugin</artifactId>
-                        <version>${surefire.version}</version>
-                        <configuration>
-                            <includes>
-                                <include>**/*ConsumerContractTestSuite.java</include>
                             </includes>
                         </configuration>
                     </plugin>
