+++ src/test/java/uk/gov/pay/ledger/pact/event/FeeIncurredEventQueueConsumerIT.java
@@ -5,12 +5,12 @@ import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
+import com.google.gson.GsonBuilder;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
 import uk.gov.pay.ledger.event.dao.EventDao;
-import uk.gov.pay.ledger.event.model.ResourceType;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
@@ -20,17 +20,19 @@ import uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture;
 
 import java.time.ZonedDateTime;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class PaymentIncludedInPayoutEventQueueContractTest {
+public class FeeIncurredEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -39,25 +41,39 @@ public class PaymentIncludedInPayoutEventQueueContractTest {
             config("queueMessageReceiverConfig.backgroundProcessingEnabled", "true")
     );
 
+    private GsonBuilder gsonBuilder = new GsonBuilder();
     private byte[] currentMessage;
     private String paymentExternalId = "payment-external-id";
-    private String payoutId = "po_12345";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
 
+    public static final long fee = 33L;
+    public static final long netAmount = 67L;
+
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentIncludedInPayoutEventPact(MessagePactBuilder builder) {
+    public MessagePact createFeeIncurredEventPact(MessagePactBuilder builder) {
+        String eventData = gsonBuilder.create()
+                .toJson(Map.of(
+                        "fee", fee,
+                        "net_amount", netAmount,
+                        "fee_breakdown", List.of(
+                                Map.of(
+                                        "amount", 24,
+                                        "fee_type", "transaction"
+                                )
+                        )
+                ));
         QueuePaymentEventFixture eventFixture = QueuePaymentEventFixture.aQueuePaymentEventFixture()
                 .withResourceExternalId(paymentExternalId)
                 .withEventDate(eventDate)
                 .withServiceId(null)
-                .withEventData(String.format("{\"gateway_payout_id\": \"%s\"}", payoutId))
-                .withEventType("PAYMENT_INCLUDED_IN_PAYOUT");
+                .withEventData(eventData)
+                .withEventType("FEE_INCURRED");
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payment included in payout message")
+                .expectsToReceive("a fee incurred message")
                 .withMetadata(metadata)
                 .withContent(eventFixture.getAsPact())
                 .toPact();
@@ -74,7 +90,6 @@ public class PaymentIncludedInPayoutEventQueueContractTest {
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(paymentExternalId).isPresent()
@@ -82,7 +97,9 @@ public class PaymentIncludedInPayoutEventQueueContractTest {
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(paymentExternalId);
         assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getGatewayPayoutId(), is(payoutId));
+        assertThat(transaction.get().getFee(), is(fee));
+        assertThat(transaction.get().getNetAmount(), is(netAmount));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"fee_breakdown\": [{\"amount\": 24, \"fee_type\": \"transaction\"}]}"));
     }
 
     public void setMessage(byte[] messageContents) {
