+++ src/test/java/uk/gov/pay/ledger/pact/event/Gateway3dsInfoObtainedEventQueueConsumerIT.java
@@ -28,7 +28,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class PaymentDetailsSubmittedByAPIEventQueueContractTest {
+public class Gateway3dsInfoObtainedEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -38,27 +38,26 @@ public class PaymentDetailsSubmittedByAPIEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "paymentDetails_externalId";
-    private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
+    private String externalId = "three_ds_info_externalId";
+    private ZonedDateTime eventDate = ZonedDateTime.parse("2021-09-07T11:40:01.123456Z");
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createPaymentDetailsSubmittedByAPIEventPact(MessagePactBuilder builder) {
-        String paymentDetailsSubmittedByAPIEventName = "PAYMENT_DETAILS_SUBMITTED_BY_API";
-        QueuePaymentEventFixture paymentDetailsSubmitted = aQueuePaymentEventFixture()
+    public MessagePact createGateway3dsVersionObtainedEventPact(MessagePactBuilder builder) {
+        String gateway3dsInfoObtainedEvent = "GATEWAY_3DS_INFO_OBTAINED";
+        QueuePaymentEventFixture gateway3dsInfoObtained = aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withEventType(paymentDetailsSubmittedByAPIEventName)
-                .withGatewayAccountId("I0YI1")
-                .withDefaultEventDataForEventType(paymentDetailsSubmittedByAPIEventName)
+                .withEventType(gateway3dsInfoObtainedEvent)
+                .withDefaultEventDataForEventType(gateway3dsInfoObtainedEvent)
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a payment details submitted by api message")
+                .expectsToReceive("a gateway 3DS info obtained message")
                 .withMetadata(metadata)
-                .withContent(paymentDetailsSubmitted.getAsPact())
+                .withContent(gateway3dsInfoObtained.getAsPact())
                 .toPact();
     }
 
@@ -66,35 +65,18 @@ public class PaymentDetailsSubmittedByAPIEventQueueContractTest {
     @PactVerification({"connector"})
     public void test() {
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        setupTransaction(transactionDao);
 
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-                        && transactionDao.findTransactionByExternalId(externalId).get().getCardholderName() != null
-        );
+                && transactionDao.findTransactionByExternalId(externalId).get().getTransactionDetails().contains("\"version_3ds\""));
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
 
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getCardBrand(), is("visa"));
-        assertThat(transaction.get().getFirstDigitsCardNumber(), is("424242"));
-        assertThat(transaction.get().getLastDigitsCardNumber(), is("4242"));
-        assertThat(transaction.get().getCardholderName(), is("J citizen"));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"expiry_date\": \"11/21\""));
-    }
-
-    private void setupTransaction(TransactionDao transactionDao) {
-        aQueuePaymentEventFixture()
-                .withEventType("PAYMENT_CREATED")
-                .withResourceExternalId(externalId)
-                .withEventDate(eventDate.minusHours(1))
-                .insert(appRule.getSqsClient());
-        await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-        );
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"version_3ds\": \"2.1.0\""));
     }
 
     public void setMessage(byte[] messageContents) {
