+++ src/test/java/uk/gov/pay/ledger/pact/event/GatewayRequires3dsAuthorisationEventQueueConsumerIT.java
@@ -22,12 +22,14 @@ import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class UserEmailCollectedEventQueueContractTest {
+public class GatewayRequires3dsAuthorisationEventQueueConsumerIT {
+
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -37,26 +39,26 @@ public class UserEmailCollectedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "userEmailCol_externalId";
-    private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
+    private String externalId = "requires_3ds_externalId";
+    private ZonedDateTime eventDate = ZonedDateTime.parse("2021-09-09T13:40:01.123456Z");
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createUserEmailCollectedEventPact(MessagePactBuilder builder) {
-        String userEmailCollectedEvent = "USER_EMAIL_COLLECTED";
-        QueuePaymentEventFixture paymentDetailsEntered = aQueuePaymentEventFixture()
+    public MessagePact createGatewayRequires3dsAuthorisationEventPact(MessagePactBuilder builder) {
+        String gatewayRequires3dsAuthorisationEvent = "GATEWAY_REQUIRES_3DS_AUTHORISATION";
+        QueuePaymentEventFixture gatewayRequires3dsAuthorisation = aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withEventType(userEmailCollectedEvent)
-                .withDefaultEventDataForEventType(userEmailCollectedEvent)
+                .withEventType(gatewayRequires3dsAuthorisationEvent)
+                .withDefaultEventDataForEventType(gatewayRequires3dsAuthorisationEvent)
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
         return builder
-                .expectsToReceive("a user email collected message")
+                .expectsToReceive("a gateway requires 3DS authorisation message")
                 .withMetadata(metadata)
-                .withContent(paymentDetailsEntered.getAsPact())
+                .withContent(gatewayRequires3dsAuthorisation.getAsPact())
                 .toPact();
     }
 
@@ -69,14 +71,14 @@ public class UserEmailCollectedEventQueueContractTest {
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-                        && transactionDao.findTransactionByExternalId(externalId).get().getEmail() != null
-        );
+                        && transactionDao.findTransactionByExternalId(externalId).get().getTransactionDetails().contains("\"requires_3ds\""));
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
 
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getEmail(), is("test@example.org"));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"version_3ds\": \"1.2.1\""));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"requires_3ds\": true"));
     }
 
     public void setMessage(byte[] messageContents) {
