+++ src/test/java/uk/gov/pay/ledger/pact/event/PaymentCreatedEventQueueConsumerIT.java
@@ -28,7 +28,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
 
-public class PaymentCreatedEventQueueContractTest {
+public class PaymentCreatedEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
