+++ src/test/java/uk/gov/pay/ledger/pact/event/PaymentDetailsSubmittedByAPIEventQueueConsumerIT.java
@@ -4,14 +4,10 @@ import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
-import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
-import com.google.gson.Gson;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
-import uk.gov.pay.ledger.event.dao.EventDao;
-import uk.gov.pay.ledger.event.model.Event;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
@@ -26,12 +22,13 @@ import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
+import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-
-public class CaptureConfirmedEventQueueContractTest {
+public class PaymentDetailsSubmittedByAPIEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -41,58 +38,63 @@ public class CaptureConfirmedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "captureConfirm_externalId";
+    private String externalId = "paymentDetails_externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
-    private String gatewayAccountId = "gateway_account_id";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createCaptureConfirmedEventPact(MessagePactBuilder builder) {
-        String eventType = "CAPTURE_CONFIRMED";
-        QueuePaymentEventFixture captureConfirmedEvent = QueuePaymentEventFixture.aQueuePaymentEventFixture()
+    public MessagePact createPaymentDetailsSubmittedByAPIEventPact(MessagePactBuilder builder) {
+        String paymentDetailsSubmittedByAPIEventName = "PAYMENT_DETAILS_SUBMITTED_BY_API";
+        QueuePaymentEventFixture paymentDetailsSubmitted = aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withGatewayAccountId(gatewayAccountId)
-                .withEventType(eventType)
-                .withDefaultEventDataForEventType(eventType)
+                .withEventType(paymentDetailsSubmittedByAPIEventName)
+                .withGatewayAccountId("I0YI1")
+                .withDefaultEventDataForEventType(paymentDetailsSubmittedByAPIEventName)
                 .withLive(true);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
-        PactDslJsonBody pactBody = captureConfirmedEvent.getAsPact();
-
         return builder
-                .expectsToReceive("a capture confirmed message")
+                .expectsToReceive("a payment details submitted by api message")
                 .withMetadata(metadata)
-                .withContent(pactBody)
+                .withContent(paymentDetailsSubmitted.getAsPact())
                 .toPact();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
-        appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
-
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
+        setupTransaction(transactionDao);
+
+        appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+                        && transactionDao.findTransactionByExternalId(externalId).get().getCardholderName() != null
         );
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getFee(), is(5L));
-        assertThat(transaction.get().getNetAmount(), is(1069L));
+        assertThat(transaction.get().getCardBrand(), is("visa"));
+        assertThat(transaction.get().getFirstDigitsCardNumber(), is("424242"));
+        assertThat(transaction.get().getLastDigitsCardNumber(), is("4242"));
+        assertThat(transaction.get().getCardholderName(), is("J citizen"));
+        assertThat(transaction.get().getTransactionDetails(), containsString("\"expiry_date\": \"11/21\""));
+    }
 
-        Map<String, Object> transactionDetails = new Gson().fromJson(transaction.get().getTransactionDetails(), Map.class);
-        assertThat(transactionDetails.get("captured_date"), is("2018-03-12T16:25:01.123456Z"));
-
-        Event event = eventDao.getEventsByResourceExternalId(externalId).get(0);
-        Map<String, Object> eventDetails = new Gson().fromJson(event.getEventData(), Map.class);
-        assertThat(eventDetails.get("gateway_event_date"),  is("2018-03-12T16:25:01.123456Z"));
-        assertThat(eventDetails.get("captured_date"),  is("2018-03-12T16:25:01.123456Z"));
+    private void setupTransaction(TransactionDao transactionDao) {
+        aQueuePaymentEventFixture()
+                .withEventType("PAYMENT_CREATED")
+                .withResourceExternalId(externalId)
+                .withEventDate(eventDate.minusHours(1))
+                .insert(appRule.getSqsClient());
+        await().atMost(1, TimeUnit.SECONDS).until(
+                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+        );
     }
 
     public void setMessage(byte[] messageContents) {
