+++ src/test/java/uk/gov/pay/ledger/pact/event/RefundIncludedInPayoutEventQueueConsumerIT.java
@@ -4,19 +4,19 @@ import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
-import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
-import com.google.gson.Gson;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
 import uk.gov.pay.ledger.event.dao.EventDao;
-import uk.gov.pay.ledger.event.model.Event;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
-import uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture;
+import uk.gov.pay.ledger.transaction.model.TransactionType;
+import uk.gov.pay.ledger.util.DatabaseTestHelper;
+import uk.gov.pay.ledger.util.fixture.QueueRefundEventFixture;
 
 import java.time.ZonedDateTime;
 import java.util.HashMap;
@@ -29,9 +29,9 @@ import static org.awaitility.Awaitility.await;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
+import static uk.gov.pay.ledger.util.fixture.TransactionFixture.aTransactionFixture;
 
-
-public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueContractTest {
+public class RefundIncludedInPayoutEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -41,57 +41,63 @@ public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueContractTest
     );
 
     private byte[] currentMessage;
-    private String externalId = "captureConfirm_externalId";
+    private String refundExternalId = "refund-external-id";
+    private String payoutId = "po_12345";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
-    private String gatewayAccountId = "gateway_account_id";
+    private QueueRefundEventFixture refundFixture;
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createCaptureConfirmedEventPact(MessagePactBuilder builder) {
-        String eventType = "STATUS_CORRECTED_TO_CAPTURED_TO_MATCH_GATEWAY_STATUS";
-        QueuePaymentEventFixture statusCorrectedToCapturedEvent = QueuePaymentEventFixture.aQueuePaymentEventFixture()
-                .withResourceExternalId(externalId)
+    public MessagePact createRefundIncludedInPayoutEventPact(MessagePactBuilder builder) {
+        refundFixture = QueueRefundEventFixture.aQueueRefundEventFixture()
+                .withResourceExternalId(refundExternalId)
                 .withEventDate(eventDate)
-                .withGatewayAccountId(gatewayAccountId)
-                .withEventType(eventType)
-                .withDefaultEventDataForEventType(eventType);
+                .withEventData(String.format("{\"gateway_payout_id\": \"%s\"}", payoutId))
+                .withParentResourceExternalId(null)
+                .withEventType("REFUND_INCLUDED_IN_PAYOUT");
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
-        PactDslJsonBody pactBody = statusCorrectedToCapturedEvent.getAsPact();
-
         return builder
-                .expectsToReceive("a status corrected to captured event")
+                .expectsToReceive("a refund included in payout message")
                 .withMetadata(metadata)
-                .withContent(pactBody)
+                .withContent(refundFixture.getAsPact())
                 .toPact();
+    }
+
+    @Before
+    public void setUp() {
+        DatabaseTestHelper.aDatabaseTestHelper(appRule.getJdbi()).truncateAllData();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
+        String parentExternalId = "payment-id";
+        aTransactionFixture()
+                .withTransactionType(TransactionType.PAYMENT.name())
+                .withExternalId(parentExternalId)
+                .insert(appRule.getJdbi());
+        aTransactionFixture()
+                .withTransactionType(TransactionType.REFUND.name())
+                .withExternalId(refundExternalId)
+                .withParentExternalId(parentExternalId)
+                .insert(appRule.getJdbi());
+
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
 
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
         EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
 
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
+                () -> transactionDao.findTransactionByExternalId(refundExternalId)
+                        .map(transactionEntity -> transactionEntity.getGatewayPayoutId() != null)
+                        .orElse(false)
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(refundExternalId);
         assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getFee(), is(5L));
-        assertThat(transaction.get().getNetAmount(), is(1069L));
-
-        Map<String, Object> transactionDetails = new Gson().fromJson(transaction.get().getTransactionDetails(), Map.class);
-        assertThat(transactionDetails.get("captured_date"), is("2018-03-12T16:25:01.123456Z"));
-
-        Event event = eventDao.getEventsByResourceExternalId(externalId).get(0);
-        Map<String, Object> eventDetails = new Gson().fromJson(event.getEventData(), Map.class);
-        assertThat(eventDetails.get("gateway_event_date"),  is("2018-03-12T16:25:01.123456Z"));
-        assertThat(eventDetails.get("captured_date"),  is("2018-03-12T16:25:01.123456Z"));
+        assertThat(transaction.get().getGatewayPayoutId(), is(payoutId));
     }
 
     public void setMessage(byte[] messageContents) {
