+++ src/test/java/uk/gov/pay/ledger/pact/event/RefundSucceededEventQueueConsumerIT.java
@@ -4,21 +4,20 @@ import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
+import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
-import uk.gov.pay.ledger.event.dao.EventDao;
+import uk.gov.pay.ledger.event.model.ResourceType;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
 import uk.gov.pay.ledger.transaction.entity.TransactionEntity;
-import uk.gov.pay.ledger.transaction.model.TransactionType;
 import uk.gov.pay.ledger.util.DatabaseTestHelper;
 import uk.gov.pay.ledger.util.fixture.QueueRefundEventFixture;
 
-import java.time.ZonedDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
@@ -29,9 +28,10 @@ import static org.awaitility.Awaitility.await;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
+import static uk.gov.pay.ledger.util.fixture.QueueRefundEventFixture.aQueueRefundEventFixture;
 import static uk.gov.pay.ledger.util.fixture.TransactionFixture.aTransactionFixture;
 
-public class RefundIncludedInPayoutEventQueueContractTest {
+public class RefundSucceededEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -41,63 +41,53 @@ public class RefundIncludedInPayoutEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String refundExternalId = "refund-external-id";
-    private String payoutId = "po_12345";
-    private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
-    private QueueRefundEventFixture refundFixture;
 
-    @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createRefundIncludedInPayoutEventPact(MessagePactBuilder builder) {
-        refundFixture = QueueRefundEventFixture.aQueueRefundEventFixture()
-                .withResourceExternalId(refundExternalId)
-                .withEventDate(eventDate)
-                .withEventData(String.format("{\"gateway_payout_id\": \"%s\"}", payoutId))
-                .withParentResourceExternalId(null)
-                .withEventType("REFUND_INCLUDED_IN_PAYOUT");
-
-        Map<String, String> metadata = new HashMap<>();
-        metadata.put("contentType", "application/json");
-
-        return builder
-                .expectsToReceive("a refund included in payout message")
-                .withMetadata(metadata)
-                .withContent(refundFixture.getAsPact())
-                .toPact();
-    }
+    private QueueRefundEventFixture refundFixture = aQueueRefundEventFixture()
+            .withResourceType(ResourceType.REFUND)
+            .withEventType("REFUND_SUCCEEDED")
+            .withGatewayTransactionId("a_gateway_transaction_id")
+            .withDefaultEventDataForEventType("REFUND_SUCCEEDED");
 
     @Before
     public void setUp() {
         DatabaseTestHelper.aDatabaseTestHelper(appRule.getJdbi()).truncateAllData();
     }
 
+    @Pact(provider = "connector", consumer = "ledger")
+    public MessagePact createRefundSucceededEventPact(MessagePactBuilder builder) {
+        Map<String, String> metadata = new HashMap<>();
+        metadata.put("contentType", "application/json");
+
+        PactDslJsonBody pactBody = refundFixture.getAsPact();
+
+        return builder
+                .expectsToReceive("a refund succeeded message")
+                .withMetadata(metadata)
+                .withContent(pactBody)
+                .toPact();
+    }
+
     @Test
     @PactVerification({"connector"})
     public void test() {
-        String parentExternalId = "payment-id";
-        aTransactionFixture()
-                .withTransactionType(TransactionType.PAYMENT.name())
-                .withExternalId(parentExternalId)
-                .insert(appRule.getJdbi());
-        aTransactionFixture()
-                .withTransactionType(TransactionType.REFUND.name())
-                .withExternalId(refundExternalId)
-                .withParentExternalId(parentExternalId)
+        aTransactionFixture()  // adds parent payment transaction for refund to be inserted
+                .withExternalId(refundFixture.getParentResourceExternalId())
+                .withTransactionType("PAYMENT")
                 .insert(appRule.getJdbi());
 
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
-
         TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
 
         await().atMost(1, TimeUnit.SECONDS).until(
-                () -> transactionDao.findTransactionByExternalId(refundExternalId)
-                        .map(transactionEntity -> transactionEntity.getGatewayPayoutId() != null)
-                        .orElse(false)
+                () -> transactionDao.findTransactionByExternalId(refundFixture.getResourceExternalId()).isPresent()
         );
 
-        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(refundExternalId);
-        assertThat(transaction.isPresent(), is(true));
-        assertThat(transaction.get().getGatewayPayoutId(), is(payoutId));
+        Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(refundFixture.getResourceExternalId());
+
+        assertThat(transaction.get().getExternalId(), is(refundFixture.getResourceExternalId()));
+        assertThat(transaction.get().getParentExternalId(), is(refundFixture.getParentResourceExternalId()));
+        assertThat(transaction.get().getCreatedDate(),is(refundFixture.getEventDate()));
+        assertThat(transaction.get().getGatewayTransactionId(), is(refundFixture.getGatewayTransactionId()));
     }
 
     public void setMessage(byte[] messageContents) {
