+++ src/test/java/uk/gov/pay/ledger/pact/event/StatusCorrectedToCapturedToMatchGatewayStatusEventQueueConsumerIT.java
@@ -4,10 +4,14 @@ import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.MessagePactProviderRule;
 import au.com.dius.pact.consumer.Pact;
 import au.com.dius.pact.consumer.PactVerification;
+import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
 import au.com.dius.pact.model.v3.messaging.MessagePact;
+import com.google.gson.Gson;
 import org.junit.Rule;
 import org.junit.Test;
 import uk.gov.pay.ledger.app.LedgerConfig;
+import uk.gov.pay.ledger.event.dao.EventDao;
+import uk.gov.pay.ledger.event.model.Event;
 import uk.gov.pay.ledger.rule.AppWithPostgresAndSqsRule;
 import uk.gov.pay.ledger.rule.SqsTestDocker;
 import uk.gov.pay.ledger.transaction.dao.TransactionDao;
@@ -22,13 +26,12 @@ import java.util.concurrent.TimeUnit;
 
 import static io.dropwizard.testing.ConfigOverride.config;
 import static org.awaitility.Awaitility.await;
-import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.mockito.Mockito.mock;
-import static uk.gov.pay.ledger.util.fixture.QueuePaymentEventFixture.aQueuePaymentEventFixture;
 
-public class Gateway3dsExemptionResultObtainedEventQueueContractTest {
+
+public class StatusCorrectedToCapturedToMatchGatewayStatusEventQueueConsumerIT {
     @Rule
     public MessagePactProviderRule mockProvider = new MessagePactProviderRule(this);
 
@@ -38,45 +41,57 @@ public class Gateway3dsExemptionResultObtainedEventQueueContractTest {
     );
 
     private byte[] currentMessage;
-    private String externalId = "exemptionResult_externalId";
+    private String externalId = "captureConfirm_externalId";
     private ZonedDateTime eventDate = ZonedDateTime.parse("2018-03-12T16:25:01.123456Z");
+    private String gatewayAccountId = "gateway_account_id";
 
     @Pact(provider = "connector", consumer = "ledger")
-    public MessagePact createGateway3dsExemptionResultObtainedEventPact(MessagePactBuilder builder) {
-        String gateway3dsExemptionResultObtainedEvent = "GATEWAY_3DS_EXEMPTION_RESULT_OBTAINED";
-        QueuePaymentEventFixture gateway3dsExemptionResultObtained = aQueuePaymentEventFixture()
+    public MessagePact createCaptureConfirmedEventPact(MessagePactBuilder builder) {
+        String eventType = "STATUS_CORRECTED_TO_CAPTURED_TO_MATCH_GATEWAY_STATUS";
+        QueuePaymentEventFixture statusCorrectedToCapturedEvent = QueuePaymentEventFixture.aQueuePaymentEventFixture()
                 .withResourceExternalId(externalId)
                 .withEventDate(eventDate)
-                .withEventType(gateway3dsExemptionResultObtainedEvent)
-                .withDefaultEventDataForEventType(gateway3dsExemptionResultObtainedEvent)
-                .withLive(true);
+                .withGatewayAccountId(gatewayAccountId)
+                .withEventType(eventType)
+                .withDefaultEventDataForEventType(eventType);
 
         Map<String, String> metadata = new HashMap<>();
         metadata.put("contentType", "application/json");
 
+        PactDslJsonBody pactBody = statusCorrectedToCapturedEvent.getAsPact();
+
         return builder
-                .expectsToReceive("a gateway 3DS exemption result obtained message")
+                .expectsToReceive("a status corrected to captured event")
                 .withMetadata(metadata)
-                .withContent(gateway3dsExemptionResultObtained.getAsPact())
+                .withContent(pactBody)
                 .toPact();
     }
 
     @Test
     @PactVerification({"connector"})
     public void test() {
-        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
-
         appRule.getSqsClient().sendMessage(SqsTestDocker.getQueueUrl("event-queue"), new String(currentMessage));
+
+        TransactionDao transactionDao = new TransactionDao(appRule.getJdbi(), mock(LedgerConfig.class));
+        EventDao eventDao = appRule.getJdbi().onDemand(EventDao.class);
 
         await().atMost(1, TimeUnit.SECONDS).until(
                 () -> transactionDao.findTransactionByExternalId(externalId).isPresent()
-                && transactionDao.findTransactionByExternalId(externalId).get().getTransactionDetails().contains("\"exemption3ds\""));
+        );
 
         Optional<TransactionEntity> transaction = transactionDao.findTransactionByExternalId(externalId);
-
         assertThat(transaction.isPresent(), is(true));
         assertThat(transaction.get().getExternalId(), is(externalId));
-        assertThat(transaction.get().getTransactionDetails(), containsString("\"exemption3ds\": \"HONOURED\""));
+        assertThat(transaction.get().getFee(), is(5L));
+        assertThat(transaction.get().getNetAmount(), is(1069L));
+
+        Map<String, Object> transactionDetails = new Gson().fromJson(transaction.get().getTransactionDetails(), Map.class);
+        assertThat(transactionDetails.get("captured_date"), is("2018-03-12T16:25:01.123456Z"));
+
+        Event event = eventDao.getEventsByResourceExternalId(externalId).get(0);
+        Map<String, Object> eventDetails = new Gson().fromJson(event.getEventData(), Map.class);
+        assertThat(eventDetails.get("gateway_event_date"),  is("2018-03-12T16:25:01.123456Z"));
+        assertThat(eventDetails.get("captured_date"),  is("2018-03-12T16:25:01.123456Z"));
     }
 
     public void setMessage(byte[] messageContents) {
